* Fixed issues that were causing warnings when compiling 64 bit in Visual Studio 2010+ * Added README.txt to the installer/windows directory to inform others about how to make their own MADARA installers. * updated to version 0.9.48
git-svn-id: http://madara.googlecode.com/svn/trunk@701 2850746a-f1c8-11dd-b507-8b4ecd5210be

+++ include/madara/knowledge_engine/Extern_Function_Variables.h
@@ -264,7 +264,7 @@ namespace Madara
        *                      range.
        * @return              entries in the resulting vector
        **/
-      unsigned int to_vector (const std::string & subject,
+      size_t to_vector (const std::string & subject,
                               unsigned int start,
                               unsigned int end,
                               std::vector <Knowledge_Record> & target);
@@ -279,7 +279,7 @@ namespace Madara
        *                      and the Knowledge Records they correspond to
        * @return              entries in the resulting map
        **/
-      unsigned int to_map    (const std::string & subject,
+      size_t to_map    (const std::string & subject,
                        std::map <std::string, Knowledge_Record> & target);
 
     private:
