+++ include/madara/knowledge_engine/Knowledge_Base.h
@@ -126,7 +126,7 @@ namespace Madara
        * @param filename           file to write to
        * @param knowledge_key       key to read the file from
        */
-      int write_file (const std::string & knowledge_key, 
+      ssize_t write_file (const std::string & knowledge_key, 
                       const std::string & filename);
 
       /**
@@ -526,7 +526,7 @@ namespace Madara
        *                      range.
        * @return              entries in the resulting vector
        **/
-      unsigned int to_vector (const std::string & subject,
+      size_t to_vector (const std::string & subject,
                               unsigned int start,
                               unsigned int end,
                               std::vector <Knowledge_Record> & target);
@@ -541,7 +541,7 @@ namespace Madara
        *                      and the Knowledge Records they correspond to
        * @return              entries in the resulting map
        **/
-      unsigned int to_map    (const std::string & subject,
+      size_t to_map    (const std::string & subject,
                        std::map <std::string, Knowledge_Record> & target);
 
     private:
